+++ crates/core/tedge_mapper/src/core/mod.rs
@@ -2,3 +2,4 @@ pub mod converter;
 pub mod error;
 pub mod mapper;
 pub mod size_threshold;
+pub mod component;
