--- crates/extensions/tedge_mapper_core/src/lib.rs
@@ -1,3 +1,0 @@
-pub mod component;
-pub mod error;
-pub mod size_threshold;
